+++ examples/default/ios/InstabugTests/InstabugSampleTests.m
@@ -8,7 +8,6 @@
 #import <XCTest/XCTest.h>
 #import "OCMock/OCMock.h"
 #import "Instabug/Instabug.h"
-#import "Instabug/IBGSurvey.h"
 #import "InstabugReactBridge.h"
 #import <Instabug/IBGTypes.h>
 #import "IBGConstants.h"
@@ -73,7 +72,7 @@
   NSArray *invocationEvents = [NSArray arrayWithObjects:[NSNumber numberWithInteger:floatingButtonInvocationEvent], nil];
   BOOL useNativeNetworkInterception = YES;
   IBGSDKDebugLogsLevel sdkDebugLogsLevel = IBGSDKDebugLogsLevelDebug;
-  
+
   OCMStub([mock setCodePushVersion:codePushVersion]);
 
   [self.instabugBridge init:appToken invocationEvents:invocationEvents debugLogsLevel:sdkDebugLogsLevel useNativeNetworkInterception:useNativeNetworkInterception codePushVersion:codePushVersion];
@@ -85,9 +84,9 @@
 - (void)testSetCodePushVersion {
   id mock = OCMClassMock([Instabug class]);
   NSString *codePushVersion = @"123";
-  
+
   [self.instabugBridge setCodePushVersion:codePushVersion];
-  
+
   OCMVerify([mock setCodePushVersion:codePushVersion]);
 }
 
@@ -234,14 +233,26 @@
 - (void)testSetReproStepsConfig {
   id mock = OCMClassMock([Instabug class]);
   IBGUserStepsMode bugMode = IBGUserStepsModeDisable;
-  IBGUserStepsMode crashMode = IBGUserStepsModeEnable;
-  IBGUserStepsMode sessionReplayMode = IBGUserStepsModeEnabledWithNoScreenshots;
+    IBGUserStepsMode anr = IBGUserStepsModeEnabledWithNoScreenshots;
 
-  [self.instabugBridge setReproStepsConfig:bugMode :crashMode :sessionReplayMode];
+  IBGUserStepsMode sessionReplayMode = IBGUserStepsModeEnabledWithNoScreenshots;
+  IBGUserStepsMode appHangsMode = IBGUserStepsModeEnabledWithNoScreenshots;
+  IBGUserStepsMode crashFatalMode = IBGUserStepsModeEnabledWithNoScreenshots;
+  IBGUserStepsMode crashNonFatalMode = IBGUserStepsModeEnable;
+  IBGUserStepsMode oomMode = IBGUserStepsModeEnable;
+  IBGUserStepsMode forceRestartMode = IBGUserStepsModeEnabledWithNoScreenshots;
+
+
+  [self.instabugBridge setReproStepsConfig:bugMode :sessionReplayMode :anr :appHangsMode : crashFatalMode :crashNonFatalMode :forceRestartMode :oomMode];
 
   OCMVerify([mock setReproStepsFor:IBGIssueTypeBug withMode:bugMode]);
-  OCMVerify([mock setReproStepsFor:IBGIssueTypeCrash withMode:crashMode]);
  OCMVerify([mock setReproStepsFor:IBGIssueTypeSessionReplay withMode:sessionReplayMode]);
+  OCMVerify([mock setReproStepsFor:IBGIssueTypeAppHang withMode:appHangsMode]);
+  OCMVerify([mock setReproStepsFor:IBGIssueTypeFatal withMode:crashFatalMode]);
+  OCMVerify([mock setReproStepsFor:IBGIssueTypeNonFatal withMode:crashNonFatalMode]);
+  OCMVerify([mock setReproStepsFor:IBGIssueTypeForceRestart withMode:forceRestartMode]);
+  OCMVerify([mock setReproStepsFor:IBGIssueTypeOutOfMemory withMode:oomMode]);
+
 }
 
 - (void)testSetUserAttribute {
@@ -448,17 +459,17 @@
   NSString* date = @"1/2/2024";
   NSInteger totalRequestCount = 10;
   NSInteger failureCount = 8;
-  
+
   NSDictionary *expected = @{
       @"date": date,
       @"totalRequestCount": @(totalRequestCount),
       @"failureCount": @(failureCount)
   };
-  
+
   OCMStub([mInstabug setWillSendNetworkDiagnosticsHandler:([OCMArg invokeBlockWithArgs:date, OCMOCK_VALUE(totalRequestCount), OCMOCK_VALUE(failureCount), nil])]);
 
   OCMStub([self.instabugBridge sendEventWithName:[OCMArg any] body:[OCMArg any]]);
-  
+
   [self.instabugBridge setOnNetworkDiagnosticsHandler];
 
   OCMVerify([self.instabugBridge sendEventWithName:@"IBGNetworkDiagnosticsHandler" body:expected]);