+++ test/modules/Instabug.spec.ts
@@ -555,14 +555,38 @@ describe('Instabug Module', () => {
     Platform.OS = 'android';
 
     const bug = ReproStepsMode.disabled;
-    const crash = ReproStepsMode.enabled;
     const sessionReplay = ReproStepsMode.enabledWithNoScreenshots;
-    const config = { bug, crash, sessionReplay };
+    const anr = ReproStepsMode.enabledWithNoScreenshots;
+    const appHangs = ReproStepsMode.enabledWithNoScreenshots;
+    const fatalCrash = ReproStepsMode.enabledWithNoScreenshots;
+    const nonFatalCrash = ReproStepsMode.enabledWithNoScreenshots;
+    const forceRestart = ReproStepsMode.enabledWithNoScreenshots;
+    const outOfMemory = ReproStepsMode.enabledWithNoScreenshots;
+
+    const config = {
+      bug,
+      sessionReplay,
+      anr,
+      appHangs,
+      fatalCrash,
+      nonFatalCrash,
+      forceRestart,
+      outOfMemory,
+    };
 
     Instabug.setReproStepsConfig(config);
 
     expect(NativeInstabug.setReproStepsConfig).toBeCalledTimes(1);
-    expect(NativeInstabug.setReproStepsConfig).toBeCalledWith(bug, crash, sessionReplay);
+    expect(NativeInstabug.setReproStepsConfig).toBeCalledWith(
+      bug,
+      sessionReplay,
+      anr,
+      appHangs,
+      fatalCrash,
+      nonFatalCrash,
+      forceRestart,
+      outOfMemory,
+    );
   });
 
   it('setReproStepsConfig should prioritize `all` over `bug`, `crash`, and `sessionReplay`', () => {
@@ -572,15 +596,41 @@ describe('Instabug Module', () => {
     const crash = ReproStepsMode.enabled;
     const sessionReplay = ReproStepsMode.enabledWithNoScreenshots;
     const all = ReproStepsMode.enabledWithNoScreenshots;
-    const config = { all, bug, crash, sessionReplay };
+    const anr = ReproStepsMode.enabledWithNoScreenshots;
+    const appHangs = ReproStepsMode.enabledWithNoScreenshots;
+    const fatalCrash = ReproStepsMode.enabledWithNoScreenshots;
+    const nonFatalCrash = ReproStepsMode.enabledWithNoScreenshots;
+    const forceRestart = ReproStepsMode.enabledWithNoScreenshots;
+    const outOfMemory = ReproStepsMode.enabledWithNoScreenshots;
+    const config = {
+      all,
+      bug,
+      crash,
+      sessionReplay,
+      anr,
+      appHangs,
+      fatalCrash,
+      nonFatalCrash,
+      forceRestart,
+      outOfMemory,
+    };
 
     Instabug.setReproStepsConfig(config);
 
     expect(NativeInstabug.setReproStepsConfig).toBeCalledTimes(1);
-    expect(NativeInstabug.setReproStepsConfig).toBeCalledWith(all, all, all);
+    expect(NativeInstabug.setReproStepsConfig).toBeCalledWith(
+      all,
+      all,
+      all,
+      all,
+      all,
+      all,
+      all,
+      all,
+    );
   });
 
-  it('setReproStepsConfig should use defaults for `bug`, `crash`, and `sessionReplay`', () => {
+  it('setReproStepsConfig should use defaults for `bug`, `crash`, `sessionReplay`, `appHangs`,`fatalCrash`,`nonFatalCrash`,`forceRestart`, and `outOfMemory`', () => {
     Platform.OS = 'android';
 
     const config = {};
@@ -590,8 +640,13 @@ describe('Instabug Module', () => {
     expect(NativeInstabug.setReproStepsConfig).toBeCalledTimes(1);
     expect(NativeInstabug.setReproStepsConfig).toBeCalledWith(
       ReproStepsMode.enabled,
-      ReproStepsMode.enabledWithNoScreenshots,
       ReproStepsMode.enabled,
+      ReproStepsMode.enabledWithNoScreenshots,
+      ReproStepsMode.enabledWithNoScreenshots,
+      ReproStepsMode.enabledWithNoScreenshots,
+      ReproStepsMode.enabledWithNoScreenshots,
+      ReproStepsMode.enabledWithNoScreenshots,
+      ReproStepsMode.enabledWithNoScreenshots,
     );
   });